# webui/openapi.yaml
openapi: 3.1.0
info:
  title: KToolBox WebUI API
  description: Authenticated REST API for the KToolBox WebUI. Selected operations also define the curated MCP tool surface
    through x-ktoolbox-mcp metadata.
  version: 1.0.0
  license:
    name: BSD-3-Clause
    identifier: BSD-3-Clause
paths:
  /api/v1/config/schema:
    get:
      summary: Get configuration schema
      operationId: config_schema
      parameters:
      - name: locale
        in: query
        required: false
        schema:
          enum:
          - zh-CN
          - zh-Hant
          - en
          - ja
          - ko
          - fr
          - ru
          type: string
          default: en
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigSchemaResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Return localized field metadata and redacted effective values for the project configuration.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: config_schema
        scope: mcp:read
        safety: read
        openWorld: false
  /api/v1/config/dotenv/{name}:
    get:
      summary: Read an environment file
      operationId: get_dotenv
      parameters:
      - name: name
        in: path
        required: true
        schema:
          enum:
          - dotenv
          - production
          type: string
          title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextDocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Read the raw project .env or prod.env document together with its revision.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: get_dotenv
        scope: null
        safety: null
        openWorld: false
    put:
      summary: Replace an environment file
      operationId: replace_dotenv
      parameters:
      - name: name
        in: path
        required: true
        schema:
          enum:
          - dotenv
          - production
          type: string
          title: Name
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextDocumentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextDocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Atomically replace the raw project .env or prod.env document after an ETag check.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: replace_dotenv
        scope: null
        safety: null
        openWorld: false
    patch:
      summary: Patch environment values
      operationId: patch_dotenv
      parameters:
      - name: name
        in: path
        required: true
        schema:
          enum:
          - dotenv
          - production
          type: string
          title: Name
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DotenvPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextDocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Atomically update selected values in .env or prod.env after an ETag check.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: patch_dotenv
        scope: null
        safety: null
        openWorld: false
  /api/v1/config/dotenv/{name}/validate:
    post:
      summary: Validate an environment file
      operationId: validate_dotenv
      parameters:
      - name: name
        in: path
        required: true
        schema:
          enum:
          - dotenv
          - production
          type: string
          title: Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextDocumentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Validate environment-file text without saving it.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: validate_dotenv
        scope: null
        safety: null
        openWorld: false
  /api/v1/config/example:
    get:
      summary: Download example configuration
      operationId: example_configuration
      responses:
        '200':
          description: Successful Response
          content:
            text/plain:
              schema:
                type: string
      description: Render and download a documented example environment file.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: example_configuration
        scope: null
        safety: null
        openWorld: false
  /api/v1/config/project:
    get:
      summary: Read project configuration
      operationId: get_project_config
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDocumentResponse'
      description: Read the raw ktoolbox.toml document and its parsed project configuration.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: get_project_config
        scope: null
        safety: null
        openWorld: false
    put:
      summary: Replace project configuration
      operationId: replace_project_config
      parameters:
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextDocumentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Atomically replace ktoolbox.toml after validation and an ETag check.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: replace_project_config
        scope: null
        safety: null
        openWorld: false
  /api/v1/config/project/validate:
    post:
      summary: Validate project configuration
      operationId: validate_project_config
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextDocumentUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Validate ktoolbox.toml text without saving it.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: validate_project_config
        scope: null
        safety: null
        openWorld: false
  /api/v1/creators:
    get:
      summary: List configured creators
      operationId: list_creators
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CreatorRosterItemResponse'
                type: array
                title: Response List Creators
      description: List the project creator roster with cached Pawchive presentation names.
      tags:
      - creators
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: list_creators
        scope: mcp:read
        safety: read
        openWorld: false
    post:
      summary: Add a creator
      operationId: add_creator
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatorReference'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatorReference'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Add a creator identity to the project roster.
      tags:
      - creators
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: add_creator
        scope: mcp:write
        safety: write
        openWorld: false
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
  /api/v1/creators/{service}/{creator_id}:
    put:
      summary: Update a creator
      operationId: update_creator
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
          title: Service
      - name: creator_id
        in: path
        required: true
        schema:
          type: string
          title: Creator Id
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatorUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatorReference'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Update the note and enabled state of a configured creator.
      tags:
      - creators
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: update_creator
        scope: mcp:write
        safety: write
        openWorld: false
    delete:
      summary: Remove a creator
      operationId: delete_creator
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
          title: Service
      - name: creator_id
        in: path
        required: true
        schema:
          type: string
          title: Creator Id
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatorReference'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Remove a creator from the project roster and clear its cached presentation profile.
      tags:
      - creators
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: delete_creator
        scope: mcp:write
        safety: destructive
        openWorld: false
  /api/v1/blockers:
    get:
      summary: List ignore rules
      operationId: list_blockers
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockerListResponse'
      description: Return the ordered project ignore-rule definitions.
      tags:
      - ignore-rules
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: list_blockers
        scope: mcp:read
        safety: read
        openWorld: false
    put:
      summary: Replace ignore rules
      operationId: replace_blockers
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/BlockerSpec'
              type: array
              title: Blockers
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockerListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Validate and replace the complete ordered ignore-rule collection.
      tags:
      - ignore-rules
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: replace_blockers
        scope: mcp:write
        safety: destructive
        openWorld: false
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
...
  /api/v1/pawchive/creators:
    get:
      summary: Search Pawchive creators
      operationId: creators
      parameters:
      - name: creator_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Creator Id
      - name: name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      - name: service
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Service
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreatorSearchItemResponse'
                title: Response Creators
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Search Pawchive for creators using an ID, name, or platform filter.
      tags:
      - pawchive
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: search_creators
        scope: mcp:read
        safety: read
        openWorld: true
...
  /api/v1/project:
    get:
      summary: Get project summary
      operationId: project
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummaryResponse'
      description: Return the active project root, configuration files, and KToolBox version.
      tags:
      - project
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: get_project_summary
        scope: mcp:read
        safety: read
        openWorld: false
...
components:
  schemas:
...
    ConfigSchemaResponse:
      properties:
        locale:
          type: string
          enum:
          - zh-CN
          - zh-Hant
          - en
          - ja
          - ko
          - fr
          - ru
          title: Locale
        sections:
          additionalProperties:
            type: string
          type: object
          title: Sections
        fields:
          items:
            $ref: '#/components/schemas/ConfigFieldResponse'
          type: array
          title: Fields
      type: object
      required:
      - locale
      - sections
      - fields
      title: ConfigSchemaResponse
    CreatorReference:
      properties:
        service:
          type: string
          minLength: 1
          title: Service
        creator_id:
          type: string
          minLength: 1
          title: Creator Id
        alias:
          anyOf:
          - type: string
          - type: 'null'
          title: Alias
        enabled:
          type: boolean
          title: Enabled
          default: true
      additionalProperties: false
      type: object
      required:
      - service
      - creator_id
      title: CreatorReference
      description: A stable Pawchive creator identity stored in the project roster.
    CreatorRosterItemResponse:
      properties:
        service:
          type: string
          minLength: 1
          title: Service
        creator_id:
          type: string
          minLength: 1
          title: Creator Id
        alias:
          anyOf:
          - type: string
          - type: 'null'
          title: Alias
        enabled:
          type: boolean
          title: Enabled
          default: true
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        avatar:
          $ref: '#/components/schemas/MediaAssetResponse'
        banner:
          $ref: '#/components/schemas/MediaAssetResponse'
      additionalProperties: false
      type: object
      required:
      - service
      - creator_id
      - avatar
      - banner
      title: CreatorRosterItemResponse
...
    DotenvPatchRequest:
      properties:
        values:
          additionalProperties:
            anyOf:
            - type: string
            - type: 'null'
          type: object
          title: Values
      type: object
      required:
      - values
      title: DotenvPatchRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProjectDocumentResponse:
      properties:
        path:
          type: string
          format: path
          title: Path
        content:
          type: string
          title: Content
        revision:
          type: string
          title: Revision
        configuration:
          $ref: '#/components/schemas/ProjectConfiguration'
      type: object
      required:
      - path
      - content
      - revision
      - configuration
      title: ProjectDocumentResponse
...
    TextDocumentResponse:
      properties:
        name:
          type: string
          title: Name
        path:
          type: string
          format: path
          title: Path
        content:
          type: string
          title: Content
        revision:
          type: string
          title: Revision
      type: object
      required:
      - name
      - path
      - content
      - revision
      title: TextDocumentResponse
    TextDocumentUpdate:
      properties:
        content:
          type: string
          title: Content
      type: object
      required:
      - content
      title: TextDocumentUpdate
...
    ValidationResponse:
      properties:
        valid:
          type: boolean
          title: Valid
          default: true
      type: object
      title: ValidationResponse
...
  securitySchemes:
    WebUISession:
      type: apiKey
      in: cookie
      name: ktoolbox_session
      description: Opaque browser session created by the login operation.
...